Gruntfile: run web app instead of grunt-connect
Rework `grunt serve` to start the full app instead of calling
grunt-connect.

Modify `web` to use relative paths to other components, instead
of depending on NODE_PATH. The reason is that NODE_PATH cannot be
modified from inside the Node process and the grunt task is running
the web app inside the same process.

Conditionally add `connect-livereload` middleware to the web app
to support the `livereload` option of the `grunt watch` task.